no need for a creator class anymore
